finances 3.7.5
The new update improves the currency parser. Now currencies (9 main ones) are updated automatically. Also, currencies that are not the main ones are also updated, but they can be updated depending on how users use them thanks to the cache system in the bot. 
The update time for currencies, stocks, and cryptocurrencies has been changed. It is now 2 hours.